Basic inotify implementation in place.
This adds a basic inotify monitor to main.c. This code still needs to be
organized and cleaned up and possibly debugged.
There is no tracking of what directory the change occured in so that
needs to be dealt with too. More or less lot's of crap left to get done.