Adds a better logger and cleans up some doc
Adds a nicer logger and commandline flags for log level. This also
introduced release and debug build options. Default debug. A library
should probably still be split out for the actual file monitoring as
well as a cleaner abort feature. Currently the program needs to be
killed through 'ctrl-c'.

READM.md has been improved with a nicer presentation of the software and
manfiles have been updated with the new flags.