Replace the -p option with positional arguments.
The -p flag was slightly cumbersome to type out when the goal was to
monitor multiple directories or files. It didn't operate will in
combination with globbing or sub commands either.

Paths should now be supplied as positional arguments instead of using
flags. This allows for pattern expansions (glob) and subcommands to be
used for a more fluid and scriptable experience.

Eg.
    mon -c <cmd> ./*.c
    mon -c <cmd> $(find . -iname *.h -o -iname *.c)